OpenSUSE 15.5: set working mirrors for zypper
OpenSUSE 15.5 provision can fail to a failed zypper install with
qemu-guest-agent due to a broken/expired mirror. This may be
eventually fixed by the package manager. This changes adds a working
source for zypper and the package.

Sometimes installing any package with zypper can output:
"Repository 'Update repository of openSUSE Backports' is invalid.".
This may mean a current mirror/repo for it is not functioning at that time. Specifying a functioning mirror fixes this.

Change-Id: I7b8ea8f1db291a1001a7d062a3da91ded99301f2
Reviewed-by: Simo Fält <simo.falt@qt.io>