CMake: Improve qt5.git "QtSynchronizeRepo.cmake" / sync-to script
The QtSynchronizeRepo.cmake script, which is meant to be driven by a
git-sync-to script, has been rewritten to support more use cases, as
well as improve the code readability and error reporting.

The script now supports the following additional use cases:
 - Clone a specified qt/{repo} submodule from code.qt.io into the
   current directory, and initialize (clone) its dependencies

- Initialize a submodule and its dependencies in an existing qt5.git
  super repo. This is similar to what init-repository does, except
  instead of using the qt5.git sha1s, it uses the dependencies.yaml
  of the specified submodule

- Support for git fetch --depth, to allow shallow cloning of the
  specified submodule and its dependencies. This is useful for CI
  where only a specific revision needs to be checked out.

The main incentive for this change is allow cloning qttools/dev/HEAD
and its dependencies in a CI run, but it's useful for daily work as
well. At some point we should check what can be merged together with
the existing init-repository script.

Pick-to: 6.8
Task-number: QTBUG-128730
Change-Id: Ie6d49d253223cc93b8831ef41d25e0adeac39b8b
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>