Add pin cores feature for all targets running tests
The feature enables 1:1 pinning of CPU cores for VM, which should mean
more stable CPU allocation from host and fewer sporadic failure.

The down side of this is that the targets will effectively use double
the capacity from the host that they previously would. It is to be seen
how effective this is in reducing flakiness, which would counter the
increased usage by reducing staging needed.

In Coin the feature is limited so that it only affects test VMs.

Task-number: QTQAINFRA-6702
Change-Id: Ie0abd68e0ce80008776f3d0fe931d996f6eb3a61
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>