Add Qt Gradle Plugin platform config
We want automated testing for QtGP. Test cases that can be run with
Gradle have been made a while ago, but they were only run locally by the
developers.

This commit adds Linux/macOS/Windows configurations for QtGP. These are
marked TestOnly as the build/test process in Gradle is packaged together
and we don't want to initiate builds for the Qt side.

Task-number: QTTA-217
Change-Id: I777d191596f43089d0ad80cbb6470c62a0eee257
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>