Extend usage of pin cores feature for all targets running tests
Extending feature usage with newest and recently added CI platforms
  - openSUSE 15.6
  - SLES 15 SP6

The feature enables 1:1 pinning of CPU cores for VM, which should mean
more stable CPU allocation from host and fewer sporadic failure.

The down side of this is that the targets will effectively use double
the capacity from the host that they previously would. It is to be seen
how effective this is in reducing flakiness, which would counter the
increased usage by reducing staging needed.

In Coin the feature is limited so that it only affects test VMs.

(Amends cf237ca8dc6e4bf111218d511f1f434189fbbf01)

Task-number: QTQAINFRA-6702
Pick-to: 6.9
Change-Id: I34f76b127899a878ef2ddf8f3c09904d8e21f493
Reviewed-by: Heikki Halmet <heikki.halmet@qt.io>