CMake: Make the InitRepository test work in Coin CI
Coin CI doesn't provide git info for the qt5.git repo, but only a
source archive. The InitRepository test needs the git info to work
correctly.

When we detect that the test is run in the CI, instead of using the
source archive, clone the qt5.git repo from code.qt.io instead. Then
fetch and checkout the exact tested ref.

Change-Id: I892545c083288a8d32d3f8957bfc64984d52f5fc
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>