Make qtgit build and work with bricscad in Build Tool.
Since the Ninja generator is not a multi config, we need to configure in
separate folders for release vs debug.

Solution: keep Release in 'build' and configure Debug in 'build_debug'.

The installation is done in the same folder 'install'. This should be
fine on Windows. Building both release and debug on Linux/Mac should be
very rare, so no need to fix this now. Especially because it would be
difficult to set QT6_PATH depending build type.

refs RM-62061